Move Synthesiser files to synthesiser subdir
